requirements.txt
djangorestframework==3.3.1